Rollout buffers: observation dtype from the space, not hard-coded uint8
RolloutBuffer, Q_RolloutBuffer and AdvRolloutBuffer all allocated observations as
np.uint8. Correct for images ([0,255] uint8), catastrophic for anything else: a
float observation in [0,1] TRUNCATED TO ZERO on write, so training read an
all-zeros tensor while collection had used the real one.

Silent, and it degraded gracefully enough to hide for a long time. The symptom
was the ADVERSARY's first-minibatch KL -- which is mathematically required to be
0 before any gradient step -- drifting 0.00000 -> 0.137, crossing
1.5*target_kl=0.075 at ~86k steps, and from ~160k aborting every ego and adv
pass before a single optimizer step. Self-play then collapsed to score_rollout=0
at a bit-reproducible 159,744 steps.

Invisible early because a near-uniform policy assigns every action ~log(1/22),
so log_prob(zeros) ~ log_prob(real). The gap only opens as the policy sharpens,
which is why it surfaced on masked RAM (which learns) and not on full RAM (whose
96.8%-constant input left the policy nearly static).

Ruled out first, each by direct measurement: the forward path (evaluate_adv_actions
reproduced the stored log-prob EXACTLY at collection, 352 samples through 209k
steps), optimizer parameter sharing (disjoint sets), side_flag/buf_num mismatch
(identical discrepancy under both argument sets), and adversary parameter drift
(fingerprints equal at collection and training).

test_buffer_alignment.py is the test that cleared the buffer's (obs, ego_action,
adv_action, log_prob) pairing -- 5/5 -- and by eliminating alignment pointed at
the dtype. It passes uint8 codes under 256, which is exactly why it went straight
through the truncation and stayed green.

Verified: float RAM and uint8 image observations both round-trip exactly.

+++ stable_baselines3/stable_baselines3/common/buffers.py
@@ -555,7 +555,16 @@ class RolloutBuffer(BaseBuffer):
 
     def reset(self) -> None:
 
-        self.observations = np.zeros((self.buffer_size, self.n_envs) + self.obs_shape, dtype=np.uint8)
+  # dtype from the observation SPACE, not hard-coded uint8. Image obs are
+        # genuinely uint8 in [0,255] so this is a no-op for them, but a float
+        # observation (--obs_type ram / info, values in [0,1]) was TRUNCATED TO
+        # ZERO on write: training then read an all-zeros observation while
+        # collection had used the real one. Symptom was the adversary's
+        # first-minibatch KL drifting 0 -> 0.14 (invisible while the policy is
+        # near-uniform, growing as it sharpens), tripping target_kl every
+        # iteration and collapsing self-play at a reproducible 159,744 steps.
+        self.observations = np.zeros((self.buffer_size, self.n_envs) + self.obs_shape,
+                                     dtype=self.observation_space.dtype)
         self.actions = np.zeros((self.buffer_size, self.n_envs, self.action_dim), dtype=np.float32)
         self.rewards = np.zeros((self.buffer_size, self.n_envs), dtype=np.float32)
         self.returns = np.zeros((self.buffer_size, self.n_envs), dtype=np.float32)
@@ -732,8 +741,18 @@ class Q_RolloutBuffer(RolloutBuffer):
     def __init__(self, buffer_size: int, observation_space: spaces.Space, action_space: spaces.Space, device: Union[th.device, str] = "auto", gae_lambda: float = 1, gamma: float = 0.99, n_envs: int = 1):
         super().__init__(buffer_size, observation_space, action_space, device, gae_lambda, gamma, n_envs)
     def reset(self) -> None:
-        self.next_observations = np.zeros((self.buffer_size, self.n_envs) + self.obs_shape, dtype=np.uint8)
-        self.observations = np.zeros((self.buffer_size, self.n_envs) + self.obs_shape, dtype=np.uint8)
+        self.next_observations = np.zeros((self.buffer_size, self.n_envs) + self.obs_shape,
+                                          dtype=self.observation_space.dtype)
+  # dtype from the observation SPACE, not hard-coded uint8. Image obs are
+        # genuinely uint8 in [0,255] so this is a no-op for them, but a float
+        # observation (--obs_type ram / info, values in [0,1]) was TRUNCATED TO
+        # ZERO on write: training then read an all-zeros observation while
+        # collection had used the real one. Symptom was the adversary's
+        # first-minibatch KL drifting 0 -> 0.14 (invisible while the policy is
+        # near-uniform, growing as it sharpens), tripping target_kl every
+        # iteration and collapsing self-play at a reproducible 159,744 steps.
+        self.observations = np.zeros((self.buffer_size, self.n_envs) + self.obs_shape,
+                                     dtype=self.observation_space.dtype)
         self.ego_actions = np.zeros((self.buffer_size, self.n_envs, self.action_dim), dtype=np.float32)
         self.adv_actions = np.zeros((self.buffer_size, self.n_envs, self.action_dim), dtype=np.float32)
         self.rewards = np.zeros((self.buffer_size, self.n_envs), dtype=np.float32)
@@ -950,7 +969,7 @@ class AdvRolloutBuffer(BaseBuffer):
             self.dones = th.zeros((self.buffer_size, self.n_envs), dtype=th.float32).pin_memory()
             self.side_flags = th.zeros((self.buffer_size, self.n_envs, 1), dtype=th.float32).pin_memory()
         else:
-            self.observations = np.zeros(obs_shape, dtype=np.uint8)
+            self.observations = np.zeros(obs_shape, dtype=self.observation_space.dtype)
             self.actions = np.zeros((self.buffer_size, self.n_envs, self.action_dim), dtype=np.float32)
             self.dstb_actions = np.zeros((self.buffer_size, self.n_envs, self.dstb_action_dim), dtype=np.float32)
             self.rewards = np.zeros((self.buffer_size, self.n_envs), dtype=np.float32)
